refactor: [M3-6388] - MUI v5 - Components > SelectRegionPanel
