--- packages/manager/src/components/SelectRegionPanel/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SelectRegionPanel';
